Preliminary work - fixed some issues with build, migrate existing CS gateway to new location, add Stripe library
