+++ Dockerfile
@@ -10,7 +10,7 @@ RUN apt-get update &&\
     apt-get install -y curl git build-essential pkg-config gdb lcov pkg-config \
       libbz2-dev libffi-dev libgdbm-dev libgdbm-compat-dev liblzma-dev \
       libncurses5-dev libreadline6-dev libsqlite3-dev libssl-dev \
-      lzma lzma-dev tk-dev uuid-dev zlib1g-dev &&\
+      lzma tk-dev uuid-dev zlib1g-dev &&\
     apt-get clean
 
 WORKDIR /tmp
