+++ apt.txt
@@ -1,1 +1,3 @@
 libxml2-dev
+libxmlsec1-dev
+libxmlsec1-openssl1
